Refactor (Two)Value(s)ToCell(s)NodeFactory
