+++ org.knime.knip.base/src/org/knime/knip/base/node/ImgPlusToImgPlusNodeFactory.java
@@ -53,7 +53,7 @@ import net.imglib2.type.numeric.RealType;
 import org.knime.knip.base.data.img.ImgPlusValue;
 import org.knime.knip.base.node.dialog.DescriptionHelper;
 import org.knime.knip.base.node.dialog.DialogComponentDimSelection;
-import org.knime.node2012.KnimeNodeDocument.KnimeNode;
+import org.knime.node.v210.KnimeNodeDocument.KnimeNode;
 
 /**
  * {@link ImgPlusToImgPlusNodeFactory}
