+++ org.knime.knip.base/src/org/knime/knip/base/node/dialog/DialogComponentOutOfBoundsSelection.java
@@ -51,9 +51,9 @@ package org.knime.knip.base.node.dialog;
 import org.knime.core.node.defaultnodesettings.SettingsModel;
 import org.knime.knip.core.types.OutOfBoundsStrategyEnum;
 import org.knime.knip.core.util.EnumUtils;
-import org.knime.node2012.ADocument.A;
-import org.knime.node2012.OptionDocument.Option;
-import org.knime.node2012.PDocument.P;
+import org.knime.node.v210.ADocument.A;
+import org.knime.node.v210.OptionDocument.Option;
+import org.knime.node.v210.PDocument.P;
 
 /**
  *
